The Controler now provides router objects on demand, the "ReRoute"-strategy uses the router provided by the Controler.
Routers are needed at different places during the simulation, e.g. in PlansPrepareForSim and also during replanning. Additionally, the exact Router can be dependent on other configuration settings, e.g. the simulation of an area-toll requires a specific router, and public transit routing will likely be scenario-dependent as well for some time. Currently, such special cases required a lot of manual configuration to get things right. Now, there is one place where the appropriate Router must be instantiated (in the Controler), and which will then be used all through the place. For compatibility reasons, the Strategies "ReRouteDijkstra" and "ReRouteLandmarks" are still availble, though the usage of the generic "ReRoute" (which uses the Controler's Router) is preferred.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@1106 a0517364-e873-4167-b1c8-416daa5a6b91