The default scoring function factory
org.matsim.core.scoring.CharyparNagelScoringFunctionFactory

now uses an instance of 

org.matsim.core.scoring.charyparNagel.CharyparNagelScoringFunctionFactory

to generate instances of scoring functions. Several classes that explicitly use plan scoring had to be adapted like the following:

- Each existing "sf.startLeg(...)" statement is now preceded with an "sf.endActivity(...)" statement.
- Each exsiting "sf.endLeg(...)" statement is now appended an "sf.startActivity(...)" statement.

This is due to the modular implementation of the CharyparNagelScoringFunction, which treats activity scoring and leg scoring in different classes, respectively. 

Furthermore, in this new implementation of the default scoring function, double arithmetics are slightly different. This may result in plan score differences after the 12th place after the decimal point compared to the existing implementation of the scoring function, which is a negligible side effect. This is why some expected plans files had to be adapted, too.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@6359 a0517364-e873-4167-b1c8-416daa5a6b91