removed knowledge from person/population:
Knowledge has now an own container called Knowledges in the package org.matsim.knowledge.*. This container class is used to store all Knowledge instances for all persons in a Map. A single Knowledge instance can be retrieved using the Id of the person.
Currently the Knowledges instance is attached to ScenarioImpl but this might change, further cleanup is needed after a discussion in the Wednesday vsp software meeting. Access to the Knowledges instance used in the Controler can currently be retrieved by a call of the deprecated method Controler.getScenarioData() and casting the Scenario to a ScenarioImpl. Don't worry about the deprecation of this method, this will be resolved.


git-svn-id: file:///srv/local/data/cleanRepo/trunk@7432 a0517364-e873-4167-b1c8-416daa5a6b91