Dear All,
I have changed a couple of things where I am not fully sure how much they affect various people.  This is in relation with the attempt to get a reasonably stable and useful API.  Unfortunately, this effort turns out to be (even) harder than expected.

The persons possibly most affected are (in decreasing probability):

Daniel Dressler
Benjamin Kickhöfer
Johannes Illenberger

For all others, I can't really say.  

The main reason is that I removed activity.setLinkId().  But it has some consequences especially in the Builders and possibly in PopulationImpl (which is a container used by nearly everybody).

Tests are running without error on my machine.

If it absolutely does not work, we may have to revert.

–––

In this context, please let me remind you of our current "contracts" that we try to fulfill.  At this point, there are two user groups:

- users.  These are people who use matsim based on config files only.  Although we do not have tagged releases, for those people it usually makes sense to look around for a nightly build that works reasonably well and then stick with it as long as possible.

- developers.  These are people who are program in Java and are inside the sourceforge repository.  The way matsim was originally set up is that we do code changes through refactorings.  The only way you can make reasonably sure that your code survives a refactoring is that you write a (reasonably fast) test case.  As an alternative, you can use some (tagged) version of matsim and try to stick with it as long as possible.  Beyond that, there is no contract that we even try to fulfill.

At this point, we are trying to set up an API so that there could be a third group:

- api-developers.  These should be people who program in Java but are not inside the sourceforge repository.  

Best

Kai

git-svn-id: file:///srv/local/data/cleanRepo/trunk@7610 a0517364-e873-4167-b1c8-416daa5a6b91