I modified the config logic of the otfvis such that it should now be possible to read config options from the config file.
The changes were as minimalistic as possible but included

- entering OTFVisConfig as a "typed" config module in the config (this should, in my understanding, NOT cause problems)

- modifying the logic in which the otf config is initialized (instead of initializing it inside otf, it is now taken from scenario).  
I modified this, however, only for the execution path that I needed for myself.  Since there are probably other execution paths, there
is no guarantee that this now works elsewhere.  

(My execution path goes through org.matsim.run, so it is the "public user interface" execution path.)

git-svn-id: file:///srv/local/data/cleanRepo/trunk@9172 a0517364-e873-4167-b1c8-416daa5a6b91