removed displayableTimeCapValue.  That was/is a nice feature, but given the many issues we have elsewhere it helps to reduce complexity a bit.
also tried to make OTFQueueSimLinkAgentsWriter non-instantiable, since it is not clear where the corresponding reader is, and without that the last chance to maintain the binary channel is gone.  Can't say if this is used anywhere since (as we know) otfvis constructs objects on the fly, but since it is a writer, there is some hope that this does not really happen.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@11340 a0517364-e873-4167-b1c8-416daa5a6b91