tried to put core.mobsim.queuesim underneath otfvis (live mode) as an alternative to QSim.  I got there to a certain extent, but things are broken:
- no display of parked vehicles
- clicking on vehicles returns the wrong vehicle
- plans are not displayed.

The main changes that I did was to replace all specific data types (QLink, QNode, QSim, QSimFeature) by data types meant for the visualizer (VisLink, VisNode, etc.).  

These are, for the time being, in otfvis.snapshots.writers, although they probably do not belong there.  As long as they are internal interfaces, this does not seem so critical.

Most of this should have been minimally invasive; apart from (manually) replacing the data types, I have done very little.  Nevertheless, I had some failing tests, which were simply due to erasing some lines that should not have been erased.  Need to cross fingers that this has not happened elsewhere where the tests don't catch it.

Dominik, I hope that the changes to not restrain you with the development of the QSim.  I think that the VisXXX interfaces should do no damage at all.  A bit more of a problem may be that I also implemented the "features" logic into the QueueSimulation ... this may bind you into a certain design.  If this is a problem, please let me know and we revert.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@11457 a0517364-e873-4167-b1c8-416daa5a6b91