Finally figured out how one can make the internal netsim engines package private -- use abstract classes instead of interfaces.  Not the greatest of all solutions, but for the time being it addresses the problem that, while interfaces themselves can be package-private, the methods that they declare are always public.
Did this for the QLink hierarchy.

Still need to do this for the SimEngine hierarchy ... but here I need to change the inherited Thread to a "composed" Thread before I can do this ... and I want a commit before that.

The QNode and the QNetwork hierarchy seem less urgent.

Tests seem to be running but there are so many commits that I have problems committing exactly what I have tested.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@12533 a0517364-e873-4167-b1c8-416daa5a6b91