renamed OTFDrawableImpl to OTFAbstractDrawableReceiver
did some stuff with lane widths so that evac vis comes out more plausible.  This amounts to no longer ignore the effectiveLaneWidth parameter from the network file.  However, the original internal value was/is NaN, which is a bit difficult to treat, so some more if conditions were necessary.

Renamed some newInstance(xxx) into newInstanceOf(xxx) since that is really what it is (in my view).

I commented out everything inside OTFQueueSimLinkAgentsWriter and made it instead extend OTFLinkAgentsHandler.Writer.  The functionality was in my view the same and now they are both using the same code.  I have left the class in place because of the connection manager in mvi files issue.

I commented out everything inside OTFLinkAgentsNoParkingHandler since the same functionality seems to be in OTFLinkAgentsHandler, and it was difficult to see which of the two ways to address this issue was used.  The 

Changed some things in OTFVisConfigGroup.  Besides some renaming and some comments, introduced an option to set the lane width (in part as a device to communicate this number from "network" to other places).  Removed the show non-parking functionality from the config file (not from the preferences) since it did not work properly; there is some additional mechanics involved.  Added "setModified" everywhere where this seemed appropriate.

Cleaned up (hopefully) a couple of places where local and global access to GL gl was intermixed.

Renamed AgentArrayDrawer into AgentArrayDrawerHelper since it is not a drawer in terms of the Drawable interface (although it is indeed functionally more of a drawer than those classes marked as Drawable).

Included an option to display network routes or not when displaying a (car) plan query.

Made methods more private in several places in otfvis.

Added comments in several places in otfvis.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@14174 a0517364-e873-4167-b1c8-416daa5a6b91