I added "VSP consistency checking" facilities.  There is
* one in the VspExperimentalConfigGroup which just checks entries from that ConfigGroup, and 

* one more general that also checks config options from other ConfigGroups.  

The latter is added in Config.java -- this is different from the "standard" config checker which is added in the controler.  I decided to do this differently since the standard config check is only added in some execution paths but not in others but I want mine to be added in all execution paths.  If you don't switch on vsp consistency checking, you should (in theory) at most get a warning.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@15695 a0517364-e873-4167-b1c8-416daa5a6b91