a couple more changes towards removing the "self-reinsert" of the agents into the mobsim.  Those changes should be switched OFF.
Also: Agents that do two activities at the same link still need an intervening leg. By current convention, this leg is a car leg with an empty route.  I moved the decision logic for this from the network engine into the higher levels of the framework (i.e. the QSim), since I don't think that we need a full-blown car departure before we decide that the agent actually stays on the same link.  -- Would be nice to find something less implicit here eventually. 

git-svn-id: file:///srv/local/data/cleanRepo/trunk@17826 a0517364-e873-4167-b1c8-416daa5a6b91