More changes towards non-self-reinserting agents.  Most of the changes are actually in the tests, which do not use the full infrastructure.
The change itself is not yet switched on.  Still, there are other changes which may affect simulation results.  As before, changes are mostly in the "dirty corners" of the simulation, such as that it is undecided if a transit driver, after she/he is done, is "aborting", or just "having an activity with end time infinity" at the last stop.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@17831 a0517364-e873-4167-b1c8-416daa5a6b91