added a config option to make "PlanSelectorForRemoval" configurable.
This is not well tested, and there is also no regression test (I don't know how to write one that makes sense).  I decided to put it in anyways since this keeps coming up, and after I looked into the code how to implement this I thought that I could as well go ahead.  This way, we might be able to give it to a student to try it out.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@18412 a0517364-e873-4167-b1c8-416daa5a6b91