changes triggered by the dreaded pt interaction activity scoring.
Main change is that this is now correctly set BOTH in the Controler AND at the location where the activity params from the config are copied to the internal object.

Did refactorings and some cleaning-up on the side.

There will probably be another commit that does some clean-up, but I need this commit in git right now in order to re-synchronize with the HEAD.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@22065 a0517364-e873-4167-b1c8-416daa5a6b91