JointPlan does not need a clique anymore
It was actually not needed anywhere, and was here just because the
Plan interface specifies a getPerson() method. A problem is that the
ExpBetaPlanSelector cannot be used in clique simulation anymore, because
it surprisingly calls the getter.
The reason for this change is that it allows to use this structure for
other groupings than "cliques" (sub-cliques in a close future, social
networks in a far one).

git-svn-id: file:///srv/local/data/cleanRepo/trunk@22091 a0517364-e873-4167-b1c8-416daa5a6b91