(Dis)utility of line switch is now considered in the standard scoring.
After some trying around, Dominik and I decided to do this by making the ScoringFunction accept PersonEvents (HasPersonId interface) directly.  This looks like it might be a bit slow, but our tests so far have not shown any speed difference.

The new mechanics is loaded by EventsToScore2 in PlanScoring; this is done to facilitate easy switching back and forth between the two versions.  If we are all happy with the new version, the previous version can go.

Also added a test that tests the (dis)utility of line switch.

While being at it, also cleaned up the pt constant: this was so far added for avery pt let while the convention is that it is added per trip.  This is now done.  (Although it is currently implemented by DEDUCTING the amounts from 2nd and onward pt let in order to remain backwards compatible with EventsToScore(2).  May also be cleaned up ...)

git-svn-id: file:///srv/local/data/cleanRepo/trunk@22210 a0517364-e873-4167-b1c8-416daa5a6b91