put Subtour analysis in TripStructureUtils
subtour analysis is now accessible in a purely static way. It uses
TripStructureUtils.getTrips(), so it is able to handle non-strict
act/leg sequences as well as trips with dummy "stage" activities,
and returns results presented in Trip form.

As a bonus point, I also find the reworked code much more readable, even
if the logic is the same.

The test hopefully covers well what was covered by
PlanAnalyzeSubtourTest, even if there are fewer testcases. It also
checks the behavior in case of complex trips and successive activities.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@23178 a0517364-e873-4167-b1c8-416daa5a6b91