subtour mode choice uses TripStructureUtils
A nice consequence is that SubtourModeChoice can thus now handle complex
trips. A potentially nasty consequence is that code that assumes that
the PlanElement instances contained by the plan are "stable" through the
run of ChooseRandomLegModeForSubtour (which was an assumption in the
test, for isntance) is now broken.

PlanAnalyzeSubtours is deprecated but not removed yet, as it is still
referenced in playgrounds (mine, to be honest).

git-svn-id: file:///srv/local/data/cleanRepo/trunk@23181 a0517364-e873-4167-b1c8-416daa5a6b91