PathSizeLogitSelector selected the selected plan as a side effect
I know this sounds silly.

In clear: it called PersonImpl.setSelectedPlan( Plan ) on the "selected"
plan before returning it. This is not what a selector is supposed to do,
though it may be unclear from the documentation.

It is particularly problematic when the selector is being used for
removal, as seems to be experimentally done...

This is the only selector where this was done so.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@24188 a0517364-e873-4167-b1c8-416daa5a6b91