MATSIM-94 returns: re-include a TimeAllocationMutator_ReRoute strategy
several users seem to desperately want this. Not sure it makes as much
sense as they think, though. See the JIRA entry for details.

Implementation note: deliberately did not include the name in the
"Names" enum. It looks dangerous, as it makes it look like you can know
the names of all strategies, while the whole idea behind the Registry
thing is that users can add easily new "named" strategies.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@25713 a0517364-e873-4167-b1c8-416daa5a6b91