commenting out broken files in "archive" package
refactoring an "archive" (which I guess it supposed to keep old unused
code) makes no sense and is even dangerous, as the kept code would not
reflect the desired state anymore. If this code is actually still used,
and you do not know how to refactor that, please contact me.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@26141 a0517364-e873-4167-b1c8-416daa5a6b91