add check that output file (does not) exist
this test passes on my machine but fails on the build server. The reason
seems to be that the file as outputed before commit 27051 is still used
on the build server, whereas it should be overriden.

This change aims at testing if the file is removed as expected on the
build server.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@27055 a0517364-e873-4167-b1c8-416daa5a6b91