do not handle 0-length car trips specially (MATSim-233)
this means that agents enter the vehicle, wait to the link and leave the
vehicle one second later. This should not change results much with
well-behaved scoring function (travel time passes from 0 to 1 sec,
assuming the vehicle is there and ready), makes the code shorter and is
more consistent (no need of the incorrect hack which triggered this
change anymore).

git-svn-id: file:///srv/local/data/cleanRepo/trunk@28569 a0517364-e873-4167-b1c8-416daa5a6b91