cache paths, not trips
based on the fact that the shortest path is not person nor departure
time dependent, even with person-dependent speeds: re-use those shortest
paths as much as possible, only updating travel times.

This is faster, but still impractical for PT...
Speedup is mainly due to bike sharing, where ODs are in limited number.
The walk part of bike sharing, on the other side, still uses more than
50% of CPU time...

git-svn-id: file:///srv/local/data/cleanRepo/trunk@29607 a0517364-e873-4167-b1c8-416daa5a6b91