drop the LRU mechanism and use a ConcurrentHashMap for softReferences
using the cache from several threads is currently very slow, because of
synchronization for each put(). ConcurrentHashMap has neat mechanisms to
reduce thread communication to the strict minimum.
The LRu was actually not very useful, given that soft references allow to
keep data almost indefinitely if enough RAM is given.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@29697 a0517364-e873-4167-b1c8-416daa5a6b91