changing the roadpricing package name back to what it was a couple of commits ago.
After some discussion, it seems that we have no clear concept of package naming in the contrib section, and also no clear argument one way or the other.  And then overall there is the argument that one should rather not change package names since that screws up everything that is not within the refactoring.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@29997 a0517364-e873-4167-b1c8-416daa5a6b91