Switching CadytsBuilder to SingleLinkMeasurement.TYPE.COUNT_VEH.
This is what it should have been all along, but the tests came with results that did not make sense.  Now they make sense.  Not fully sure why this is: We changed the getSimValue implementation some days ago (by really summing up multi-hour measurements), but it is still not clear why this should have changed the test since it contains entries only in one hour.  (Possibly, the entry is in the _second_ hour of a pair, and was thus ignored by the previous version of the test -- this is all pretty confusing since hours start counting at 1 while arrays start counting at 0.)

git-svn-id: file:///srv/local/data/cleanRepo/trunk@30168 a0517364-e873-4167-b1c8-416daa5a6b91