Separated vehicles and transitVehicles container.
Did that by

(1) renaming everything related to the old "vehicles" container to "transitVehicles".

(2) Introduced a new "vehicles" container, together with a new config section.

(3) Tried to check for all occurences of what was now scenario.getTransitVehicles() if this should have been vehicles.

I also replaced the config switch isUseVehicles that had to be "on" in the past by "isUseTransit" which normally is on anyways if you want to use pt.  (Maybe there are use cases where one would not want this, but TransitSchedule is under the same switch so at least it is not making the situation worse.)

git-svn-id: file:///srv/local/data/cleanRepo/trunk@31995 a0517364-e873-4167-b1c8-416daa5a6b91