fail if adding twice the same strategy to the same subpopulation
Had some unexpected behavior because of that in Zurich, where one of our
scenarios had some strategies defined twice, to have more complex weight
evolution from the config file. This used to work, as the factory was
called each time, and thus a new instance was passed for each setting.

git-svn-id: file:///srv/local/data/cleanRepo/trunk@32889 a0517364-e873-4167-b1c8-416daa5a6b91