hopefully fixing failing contrib tests
I think that the tests were failing because of the mode router factory change.  Looking at the plans file differences revealed that in new output first and last link of the route are included into the xml route whereas previously they were not.

Rather than just copying the new files over the old files, I wrote a converter to change the old files into the new format.  Still resulted in failure.  Reason is that in new version the walk mode is routed as a network mode.  (Not sure why that suddenly occurs.)  I checked for FastAStar manually and that were the only differences, and after that I took the new output and copied it over the old input.

Still didn't work ... one has to run maven update in order to make eclipse "see" these files.  Now it seems to be working.