MATSIM-440, MATSIM-431. PopulationReaders created wrong types of routes for some leg modes.
Detailed problem description: In plans_v4 (generally before MATSIM-431), the type of a route being read from file was configured in the code by leg-mode, i.e. in the code one could specify that a route within a car-leg should be parsed as NetworkRoute, and a route within a pt-leg should be parsed as ExperimentalTransitRoute. Naturally, this could lead to problems if one did not correctly configure this. Now in population_v5, routes are written with a "type", and when reading in a file, the correct class is determined based on that type.

2 problems existed:
- reading population_v5 had GenericRouteImpl assigned to routes of type "experimentalPT1". ExperimentalTransitRoute would have been correct.
- reading plans_v4 created NetworkRoutes for nearly everything except leg mode "pt". As this can no longer be configured in the code, the PlansV4-reader now assumes "car" legs to contain a NetworkRoute, "pt" legs to contain an ExperimentalTransitRoute, and all other legs to contain a GenericRouteImpl. It's not yet perfect, but hopefully useful for most setups.