first checkpoint after QLane, QLink refactorings
Everything compiles again.  Will now send this to travis.

General idea was to make the QLanes only depend on the QLinks, and the
QLinks only on the QNetsimEngine.  I.e. each object might have a back
pointer to the next object above, but this pointer should not be
available to the outside so that these back pointers cannot be
chained.  In order to facilitate this, some central objects such as
EventsManager, AgentCounter, MobsimTimer were made available from a
central NetsimEngineContext.