A bug fix
Two links that are belonging to the same road but go in the opposite
directions will have the same coordinate in the network. This was
creating problems when removing the vehicle upon reservation from the
quad tree. It was happening that the vehicle was not removed and then
tried to be taken from somebody else resulting in an error.