Added error check and message to deal with crashing code
If an osm-file is incomplete it may cause the code to crash without useful error messages. This commit just adds an error check with a response message that hints towards a solution so that the user can fix the core problem: not parsing the osm-file originally using the correct arguments.