Fixing bug in CoordUtils that made tests fail
The bug was of the type that in one case it used hasZ=true to go into the 2d version, and hasZ=false to go into the 3d version.

Found _two_ methods where that was the case.  The second one (rotateToRight) is not covered by a test, so the tests would not have found it.