Added a comment
KeepLastSelected strategy does not do as one might expect it to do. If
one expects to have the plan that was executed in the previous
iteration, one might be surprised. As there is a limit to plans memory,
the selected plan from the previous iteration might be removed and new
selected plan would be random.