Dijkstra: Adapted heap initialization. Now, initially all nodes of the network are put into the heap in the same order as they are in the network. The heap assigns each of them an integer as index. If during the routing two elements have the same priority, the heap orders them based on their index. Due to this new initialization process, the heap sorts them deterministic independent of the order the nodes are processed by the router (by default, the heap assigns the indices when an element is added for the first time).
Dijkstra and AStarEuclidean:
In case a node can be reached from more than one link with exactly the
same costs, it is decided based on the links' ids which one will be used
in the resulting path. So far, the first one that was processed was
used. However, as a result, the outcome depended on the order in which
the links were processed by the router.