Router debugging (#85)
* ArrayRoutingNetworkFactory: ensure only one Network at a time is
created. Otherwise, the internal indices counter would get mixed up.

* FastRouterFactories: create PreProcessData when the router instance is
created using the same network/travel time/travel disutility objects as
the router uses.

* AbstractRoutingNetwork: use a LinkedHashMap to store the nodes to ensure
that iterating over them is deterministic.

* Dijkstra:
Adapted heap initialization. Now, initially all nodes of the network are
put into the heap in the same order as they are in the network. The heap
assigns each of them an integer as index.
If during the routing two elements have the same priority, the heap
orders them based on their index. Due to this new initialization
process, the heap sorts them deterministic independent of the order the
nodes are processed by the router (by default, the heap assigns the
indices when an element is added for the first time).

Dijkstra and AStarEuclidean:
In case a node can be reached from more than one link with exactly the
same costs, it is decided based on the links' ids which one will be used
in the resulting path. So far, the first one that was processed was
used. However, as a result, the outcome depended on the order in which
the links were processed by the router.

* Possible workaround for travel time jumps at bin boarders?

* WrappedBinaryMinHeap: do not clear internal lookup map "object -> index"
anymore. In case new object should be added after reseting the heap
please create a new heap.

* Dijkstra and AStarEuclidean: use the same map so store the NodeData
objects.

* fix a bug in the linear interpolation of travel times

* Cleanup in routing code.
Switched to linear interpolated travel times in ReRoutingIT test.

* fix merge: factories were moved

* clean up 'fast' router factories: only set dead end data of nodes once

* fix ReRoutingIT (see MATSim-730)