changes around: interaction activities; old config file format
Seems that specifying _any_ activity parameter in the config file removed the interaction activities for access/egress routing.  Not sure why nobody complained, but maybe those who were using this just entered them manually into the config file.  This, however, runs the
risk that someone by accident does not use the "not scoring" setting for those, which will send the scoring function to wrong and possibly numerically unstable values.

So now these are auto-generated based on the routing config.

I also starting putting in flags and warnings to start phasing out config v1.  Advantage of this would be that PlanCalcXxxConfigGroups would become a bit less complex.