set typed event as "oneof"
This is what was implemented by hand. The docs
(https://developers.google.com/protocol-buffers/docs/proto#updating)
say:

> Changing a single optional value into a member of a new oneof is safe
> and binary compatible. Moving multiple optional fields into a new oneof
> may be safe if you are sure that no code sets more than one at a time.
> Moving any fields into an existing oneof is not safe.

We are in the second case, so that should be safe.
What surprises me is that the code required adaptations, because
checking which field is set is done differently.

As it was now redundant, I also removed the "type" field. Again, the
docs say

> Non-required fields can be removed, as long as the field number is not
> used again in your updated message type. You may want to rename the
> field instead, perhaps adding the prefix "OBSOLETE_", or make the field
> number reserved, so that future users of your .proto can't accidentally
> reuse the number.

So that should be fine.

I am a bit unsure about what type of compatibility it relates to,
however. They speak of "not breaking existing code", but here the
concern would rather be "being able to ingest old format". I understand
compatibility goes in both directions, but I am not sure.