Modify Bicycle Contrib to allow Link and Person Based Scoring
I am currently trying to modify the bicycle contrib so I can use the
results from a route choice choice modeling experiment. This means I
need to "neutralize" certain parts of the scoring functionality so that
the resulting scoring is consistant with the design of the choice
experiment. I've run into problems finding all the code that I need to
neutralize, especially with regards to the routing.