changes to make freight-dfg17 work with changes in VehicleType
I ended up bringing CarrierVehicleType back.  With this commit, there is also a VehicleType interface, CarrierVehicleType is an implementation of it, and uses VehicleTypeImpl as a delegate.

However, I will next try to go back to VehicleType as a final class.

There are related changes elsewhere in the hierarchy, especially (Vehicle)CostInformation.

Also moved deprecated methods into VehicleType so that I have all the different versions in some place (getFixedCost() vs getFix() etc.).