Fixing problem with Adaptive Real Time Rebalancing Strategy
When grid is used to divde the network into zones, there are usually a large number of  zones. Many of them are empty or contain no demand at all. Therefore, it makes sense to ignore those zones when performing the rebalance. This fix enables the strategy to ignore the zone without any demand (e.g. forest, lakes, outside the network).