Update PreviousIterationDRTDemandEstimator.java
Perhaps it is a good idea to split the time bin size with the rebalance interval. For Adaptive real time rebalancing strategy, the time bin size and the rebalance interval are unrelated. If the time bin size is too small, there will be some problem. As a  result. a temporary solution of adding a max statement is used here. We can have a disucssion for a permanant solution.