Update travel time validator (based on HERE API) (#1215)
The HERE API has gone through some changes. Therefore, we also need to update the way the request validator. 

Example on getting route from HERE can be found in this link (the url, the format and the return data format):
https://developer.here.com/documentation/routing-api/8.9.1/dev_guide/topics/use-cases/calculate-route.html 
In addition to the parameters shown in the example, we also need to add additional parameters (e.g. API key, departure time)

Note that the credentials in HERE has changed. The APP ID and APP Code is no longer provided in new projects: https://developer.here.com/documentation/authentication/dev_guide/index.html
The API key is used instead, which can be acquired from the HERE.com with a HERE account.