Enhance Hermes to support multiple transit routes with same Id
Previously, Hermes assumed that TransitRoute ids are globally unique, and used its index in various places.
As this is actually not the case, we need to enumerate all transit routes, using this number (routeNo) as reference through Hermes. Also, we have to provide access to the original TransitLine-id and TransitRoute-id in some places.

See #1248