Get zone for link ID: use to node of the link
When determining the links (i.e. DRT stops) within the service area shapefile, the coordinate of the "toNode" of the link is used (DrtModeModule, line 229), Therefore, I think it is better to use To node of the link when determin the zone a link belongs to. If we use the link's coordinate, there are likely to be some requests within the service area but outside any of the zone. This will add some unwanted effect to the analysis.