Get zone for link ID: use to node of the link (#1287)
Now we use link.getToNode.getCoord for the DRT zonal system (i.e. getZoneForLinkId) and for the analysis output (i.e. the coordinate of the request origin and destination). Now (hopefully) it should be consistent in the whole DRT pacakge. In case there is additional inconsistency that I forgot to change, please let me know. Thank you very much!