SRR performance improvement
We had a situation where every time intermodal stops are searched where
specific stop attribute is used as a filter a new quad tree was
generated with those stops. This was making intermodal router to
considerably underperform.

This commit ensures that the quad tree is created only if it does not
exist in the method prepareStopFilterQuadTreeIfNotExistent

In our set-up (KNR, PNR, walk) this leads to speedup of factor ~15 for
intermodal routing