allow for setting arrivalOffset and departureOffset as OptionalTime in TransitRouteStop
the variable type is already OptionalTime, but there is no setter accepting OptionalTime, only double. At first glance this might introduce more issues with TransitRouteStops having an undefined arrival or departure offset, but the public builder of TransitRouteStopImpl already sets those to OptionalTime.undefinedTime() by default and they remain undefined if the arrivalOffset setter of the builder is not called. Defined times for all but the first stop's arrival and the last stop's departure should be checked somewhere else, possibly during simulation startup.