integrate functionality of ExperiencedPlanModule into PlansScoringModule
This ensures that a ScoringFunction is called by only a single EventHandler and thus thread. Before it was possible for a ScoringFunction that used the arbitrary events to be called from two different threads, leading to race conditions.

Renamed ExperiencedPlansModule to StandaloneExperiencedPlansModule to reflect in the name that is is only to be used in special setups where no scoring is being used, e.g. in post-processing data.

Fixes #1593